+++ src/test/java/issue10/Eventing.java
@@ -1,5 +1,0 @@
-package issue10;
-
-public interface Eventing {
-    MetricsConfig getMetrics();
-}