+++ src/test/java/issue10/MetricsConfig.java
@@ -1,5 +1,0 @@
-package issue10;
-
-public interface MetricsConfig {
-    boolean isJvmGaugesEnabled();
-}