+++ src/main/groovy/com/github/tonybaines/gestalt/transformers/DefaultPropertyNameTransformer.groovy
@@ -1,0 +1,5 @@
+package com.github.tonybaines.gestalt.transformers
+
+class DefaultPropertyNameTransformer implements PropertyNameTransformer {
+  String fromPropertyName(String propertyName) { propertyName }
+}