+++ src/main/java/com/github/tonybaines/gestalt/transformers/PropertyNameTransformer.java
@@ -1,6 +1,5 @@
 package com.github.tonybaines.gestalt.transformers;
 
 public interface PropertyNameTransformer {
-  public abstract String fromPropertyName(String propertyName);
-  public abstract String toPropertyName(String value);
+  String fromPropertyName(String propertyName);
 }