+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # [0.8.4](PENDING)
 * Issue #12: NPE when accessing unassigned primitive values (fixed in README.md)
-* Issue #10: Converting to a Properties instance more than two levels deep fails
+* Issue #10: NPE when serialising a config instance with null/missing values to a Properties instance
 * Issue #9: Converting to a Properties instance more than two levels deep fails
 
 # [0.8.3](http://central.maven.org/maven2/com/github/tonybaines/gestalt/0.8.3) (2014-06-29)